+++ src/pins/address_bus.h
@@ -7,10 +7,10 @@
 
 #include <stdint.h>
 
-uint8_t address_bus;
+uint16_t address_bus;
 
 void initialize_address_bus();
-void write_address_bus(uint8_t byte);
-uint8_t read_address_bus();
+void write_address_bus(uint16_t byte);
+uint16_t read_address_bus();
 
 #endif //INC_8080_EMULATOR_ADDRESS_BUS_H
